+++ test/test_programs.py
@@ -1,0 +1,243 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2021.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Tests for program related runtime functions."""
+
+import copy
+import json
+import os
+from io import StringIO
+from unittest.mock import patch
+import warnings
+
+from qiskit_ibm_runtime.exceptions import IBMInputValueError
+from qiskit_ibm_runtime.exceptions import RuntimeProgramNotFound
+from qiskit_ibm_runtime.runtime_program import ParameterNamespace
+
+from .ibm_test_case import IBMTestCase
+from .mock.fake_runtime_service import FakeRuntimeService
+from .utils.program import upload_program, DEFAULT_DATA, DEFAULT_METADATA
+
+
+import unittest
+class TestPrograms(unittest.TestCase):
+    """Class for testing runtime modules."""
+
+    def setUp(self):
+        """Initial test setup."""
+        super().setUp()
+        self._legacy_service = FakeRuntimeService(auth="legacy", token="some_token")
+
+    def test_list_programs(self):
+        """Test listing programs."""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        programs = service.programs()
+        all_ids = [prog.program_id for prog in programs]
+        self.assertIn(program_id, all_ids)
+
+    def test_list_programs_with_limit_skip(self):
+        """Test listing programs with limit and skip."""
+        program_ids = []
+        service = self._legacy_service
+        for _ in range(3):
+            program_ids.append(upload_program(service))
+        programs = service.programs(limit=2, skip=1)
+        all_ids = [prog.program_id for prog in programs]
+        self.assertNotIn(program_ids[0], all_ids)
+        self.assertIn(program_ids[1], all_ids)
+        self.assertIn(program_ids[2], all_ids)
+        programs = service.programs(limit=3)
+        all_ids = [prog.program_id for prog in programs]
+        self.assertIn(program_ids[0], all_ids)
+
+    def test_list_program(self):
+        """Test listing a single program."""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        program = service.program(program_id)
+        self.assertEqual(program_id, program.program_id)
+
+    def test_print_programs(self):
+        """Test printing programs."""
+        service = self._legacy_service
+        ids = []
+        for idx in range(3):
+            ids.append(upload_program(service, name=f"name_{idx}"))
+
+        programs = service.programs()
+        with patch("sys.stdout", new=StringIO()) as mock_stdout:
+            service.pprint_programs()
+            stdout = mock_stdout.getvalue()
+            for prog in programs:
+                self.assertIn(prog.program_id, stdout)
+                self.assertIn(prog.name, stdout)
+                self.assertNotIn(str(prog.max_execution_time), stdout)
+            service.pprint_programs(detailed=True)
+            stdout_detailed = mock_stdout.getvalue()
+            for prog in programs:
+                self.assertIn(prog.program_id, stdout_detailed)
+                self.assertIn(prog.name, stdout_detailed)
+                self.assertIn(str(prog.max_execution_time), stdout_detailed)
+
+    def test_upload_program(self):
+        """Test uploading a program."""
+        service = self._legacy_service
+        max_execution_time = 3000
+        is_public = True
+        program_id = upload_program(
+            service=service, max_execution_time=max_execution_time, is_public=is_public
+        )
+        self.assertTrue(program_id)
+        program = service.program(program_id)
+        self.assertTrue(program)
+        self.assertEqual(max_execution_time, program.max_execution_time)
+        self.assertEqual(program.is_public, is_public)
+
+    def test_update_program(self):
+        """Test updating program."""
+        service = self._legacy_service
+        new_data = "def main() {foo=bar}"
+        new_metadata = copy.deepcopy(DEFAULT_METADATA)
+        new_metadata["name"] = "test_update_program"
+        new_name = "name2"
+        new_description = "some other description"
+        new_cost = DEFAULT_METADATA["max_execution_time"] + 100
+        new_spec = copy.deepcopy(DEFAULT_METADATA["spec"])
+        new_spec["backend_requirements"] = {"input_allowed": "runtime"}
+
+        sub_tests = [
+            {"data": new_data},
+            {"metadata": new_metadata},
+            {"data": new_data, "metadata": new_metadata},
+            {"metadata": new_metadata, "name": new_name},
+            {
+                "data": new_data,
+                "metadata": new_metadata,
+                "description": new_description,
+            },
+            {"max_execution_time": new_cost, "spec": new_spec},
+        ]
+
+        for new_vals in sub_tests:
+            with self.subTest(new_vals=new_vals.keys()):
+                program_id = upload_program(service)
+                service.update_program(program_id=program_id, **new_vals)
+                updated = service.program(program_id, refresh=True)
+                if "data" in new_vals:
+                    raw_program = service._api_client.program_get(program_id)
+                    self.assertEqual(new_data, raw_program["data"])
+                if "metadata" in new_vals and "name" not in new_vals:
+                    self.assertEqual(new_metadata["name"], updated.name)
+                if "name" in new_vals:
+                    self.assertEqual(new_name, updated.name)
+                if "description" in new_vals:
+                    self.assertEqual(new_description, updated.description)
+                if "max_execution_time" in new_vals:
+                    self.assertEqual(new_cost, updated.max_execution_time)
+                if "spec" in new_vals:
+                    raw_program = service._api_client.program_get(program_id)
+                    self.assertEqual(new_spec, raw_program["spec"])
+
+    def test_update_program_no_new_fields(self):
+        """Test updating a program without any new data."""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        with warnings.catch_warnings(record=True) as warn_cm:
+            service.update_program(program_id=program_id)
+            self.assertEqual(len(warn_cm), 1)
+
+    def test_delete_program(self):
+        """Test deleting program."""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        service.delete_program(program_id)
+        with self.assertRaises(RuntimeProgramNotFound):
+            service.program(program_id, refresh=True)
+
+    def test_double_delete_program(self):
+        """Test deleting a deleted program."""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        service.delete_program(program_id)
+        with self.assertRaises(RuntimeProgramNotFound):
+            service.delete_program(program_id)
+
+    def test_retrieve_program_data(self):
+        """Test retrieving program data"""
+        service = self._legacy_service
+        program_id = upload_program(service, name="qiskit-test")
+        service.programs()
+        program = service.program(program_id)
+        self.assertEqual(program.data, DEFAULT_DATA)
+        self._validate_program(program)
+
+    def test_program_params_validation(self):
+        """Test program parameters validation process"""
+        service = self._legacy_service
+        program_id = upload_program(service)
+        program = service.program(program_id)
+        params: ParameterNamespace = program.parameters()
+        params.param1 = "Hello, World"
+        # Check OK params
+        params.validate()
+        # Check OK params - contains unnecessary param
+        params.param3 = "Hello, World"
+        params.validate()
+        # Check bad params - missing required param
+        params.param1 = None
+        with self.assertRaises(IBMInputValueError):
+            params.validate()
+        params.param1 = "foo"
+
+    def test_program_metadata(self):
+        """Test program metadata."""
+        # TODO We should probably use a temp file
+        service = self._legacy_service
+        file_name = "test_metadata.json"
+        with open(file_name, "w") as file:
+            json.dump(DEFAULT_METADATA, file)
+        self.addCleanup(os.remove, file_name)
+
+        sub_tests = [file_name, DEFAULT_METADATA]
+
+        for metadata in sub_tests:
+            with self.subTest(metadata_type=type(metadata)):
+                program_id = service.upload_program(
+                    data=DEFAULT_DATA, metadata=metadata
+                )
+                program = service.program(program_id)
+                service.delete_program(program_id)
+                self._validate_program(program)
+
+    def _validate_program(self, program):
+        """Validate a program."""
+        self.assertEqual(DEFAULT_METADATA["name"], program.name)
+        self.assertEqual(DEFAULT_METADATA["description"], program.description)
+        self.assertEqual(
+            DEFAULT_METADATA["max_execution_time"], program.max_execution_time
+        )
+        self.assertTrue(program.creation_date)
+        self.assertTrue(program.update_date)
+        self.assertEqual(
+            DEFAULT_METADATA["spec"]["backend_requirements"],
+            program.backend_requirements,
+        )
+        self.assertEqual(
+            DEFAULT_METADATA["spec"]["parameters"], program.parameters().metadata
+        )
+        self.assertEqual(
+            DEFAULT_METADATA["spec"]["return_values"], program.return_values
+        )
+        self.assertEqual(
+            DEFAULT_METADATA["spec"]["interim_results"], program.interim_results
+        )
